Add context for Detective Carter
Add the context that Carter is a police detective, and that's she's
watching security footage (not the Chinese station CCTV).

Also that she's former army.

Pull the army details from a wiki page.

"02:15 Carter watches a few seconds of CCTV video of Reece brawling with
Anton’s gang to determine it’s a fighting style not taught in the
“regular army”.  Because it’s an elite close-quarters combat style that
trains you to knee your opponent in the stomach."