Add the list in the library
It's not clear to me now to express spacial information between Place
and non-Place.

* the list is on the evidence board
* the evidence board is in the library
* Finch owns the library

"13:00 Finch brings Reese to the abandoned library hideout and shows him
the Crazy Wall. We see The List, a number of social security numbers
printed on “green bar” tractor feed paper. Except the paper is much
narrower than usual fanfold list paper, and the text is in OCR-A, not
something that looks like a dot-matrix output, at a resolution under the
usual 80 columns a line printer would produce."